+++ stroom-pipeline/src/main/java/stroom/pipeline/server/reader/ReaderRecordDetector.java
@@ -46,6 +46,10 @@ public class ReaderRecordDetector extends FilterReader implements RecordDetector
 
     @Override
     public int read(final char buf[], final int off, final int len) throws IOException {
+        if (controller == null) {
+            return super.read(buf, off, len);
+        }
+
         if (end) {
             return -1;
         }