+++ stroom-pipeline/src/main/java/stroom/pipeline/server/writer/RollingStreamAppender.java
@@ -54,6 +54,12 @@ import java.io.IOException;
         PipelineElementType.ROLE_TARGET, PipelineElementType.ROLE_DESTINATION,
         PipelineElementType.VISABILITY_STEPPING}, icon = ElementIcons.STREAM)
 public class RollingStreamAppender extends AbstractRollingAppender implements RollingDestinationFactory {
+    private static final int MB = 1024 * 1024;
+    private static final int DEFAULT_ROLL_SIZE = 100 * MB;
+
+    private static final long SECOND = 1000;
+    private static final long MINUTE = 60 * SECOND;
+    private static final long HOUR = 60 * MINUTE;
 
     private final StreamStore streamStore;
     private final StreamHolder streamHolder;
@@ -65,6 +71,11 @@ public class RollingStreamAppender extends AbstractRollingAppender implements Ro
     private Feed feed;
     private String streamType;
     private boolean segmentOutput = true;
+    private Long frequency = HOUR;
+    private SimpleCron schedule;
+    private long rollSize = DEFAULT_ROLL_SIZE;
+
+    private boolean validatedSettings;
 
     private StreamKey key;
 
@@ -104,7 +115,7 @@ public class RollingStreamAppender extends AbstractRollingAppender implements Ro
     }
 
     @Override
-    Object getKey() throws IOException {
+    Object getKey() {
         if (key == null) {
             key = new StreamKey(feed, streamType, segmentOutput);
         }
@@ -128,8 +139,17 @@ public class RollingStreamAppender extends AbstractRollingAppender implements Ro
             }
         }
 
-        if (streamType == null) {
-            throw new ProcessException("Stream type not specified");
+            if (streamType == null) {
+                throw new ProcessException("Stream type not specified");
+            }
+
+            if (frequency != null && frequency <= 0) {
+                throw new ProcessException("Rolling frequency must be greater than 0");
+            }
+
+            if (rollSize <= 0) {
+                throw new ProcessException("Roll size must be greater than 0");
+            }
         }
     }
 
@@ -144,9 +164,40 @@ public class RollingStreamAppender extends AbstractRollingAppender implements Ro
         this.streamType = streamType;
     }
 
-    @PipelineProperty(description = "Shoud the output stream be marked with indexed segments to allow fast access to individual records?", defaultValue = "true")
+    @PipelineProperty(description = "Should the output stream be marked with indexed segments to allow fast access to individual records?", defaultValue = "true")
     public void setSegmentOutput(final boolean segmentOutput) {
         this.segmentOutput = segmentOutput;
+    }
+
+    @PipelineProperty(description = "Choose how frequently streams are rolled.", defaultValue = "1h")
+    public void setFrequency(final String frequency) {
+        if (frequency == null || frequency.trim().length() == 0) {
+            this.frequency = null;
+        } else {
+            try {
+                final Long value = ModelStringUtil.parseDurationString(frequency);
+                if (value == null || value <= 0) {
+                    throw new PipelineFactoryException("Incorrect value for frequency: " + frequency);
+                }
+
+                this.frequency = value;
+            } catch (final NumberFormatException e) {
+                throw new PipelineFactoryException("Incorrect value for frequency: " + frequency);
+            }
+        }
+    }
+
+    @PipelineProperty(description = "Provide a cron expression to determine when streams are rolled.")
+    public void setSchedule(final String expression) {
+        if (expression == null || expression.trim().length() == 0) {
+            this.schedule = null;
+        } else {
+            try {
+                this.schedule = SimpleCron.compile(expression);
+            } catch (final NumberFormatException e) {
+                throw new PipelineFactoryException("Incorrect value for schedule: " + expression);
+            }
+        }
     }
 
     @PipelineProperty(description = "Choose the maximum size that a stream can be before it is rolled.", defaultValue = "100M")