+++ stroom-pipeline/src/main/java/stroom/pipeline/server/writer/StreamAppender.java
@@ -212,4 +212,9 @@ public class StreamAppender extends AbstractAppender {
     public void setRollSize(final String size) {
         super.setRollSize(size);
     }
+
+    @PipelineProperty(description = "Choose if you want to split aggregated streams into separate output streams.", defaultValue = "false")
+    public void setSplitAggregatedStreams(final boolean splitAggregatedStreams) {
+        super.setSplitAggregatedStreams(splitAggregatedStreams);
+    }
 }